backend/requirements.txt
fastapi
uvicorn[standard]
openai
requests
beautifulsoup4
python-multipart
python-dotenv
pypdf
aisuite[openai]
